requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.16.0
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cohere==4.45
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
fastavro==1.9.3
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2023.12.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-aiplatform==1.40.0
google-cloud-bigquery==3.17.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.4.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.13.3
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.15.1
torch==2.1.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
transformers==4.37.1
types-protobuf==4.24.0.20240106
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
wordcloud==1.9.3
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
